use callback function instead of progress.Incrementer

+++ internal/fileutils/fileutils.go
@@ -16,7 +16,6 @@ import (
 	"github.com/ActiveState/cli/internal/failures"
 	"github.com/ActiveState/cli/internal/locale"
 	"github.com/ActiveState/cli/internal/logging"
-	"github.com/ActiveState/cli/internal/progress"
 )
 
 // FailFindInPathNotFound indicates the specified file was not found in the given path or parent directories
@@ -423,6 +422,9 @@ func IsEmptyDir(path string) (bool, *failures.Failure) {
 	return (len(files) == 0), nil
 }
 
+// MoveAllFilesCallback is invoked for every file that we move
+type MoveAllFilesCallback func()
+
 // MoveAllFilesRecursively moves files and directories from one directory to another.
 // Unlike in MoveAllFiles, the destination directory does not need to be empty, and
 // may include directories that are moved from the source directory.
@@ -430,7 +432,7 @@ func IsEmptyDir(path string) (bool, *failures.Failure) {
 // Warnings are printed if
 // - a source file overwrites an existing destination file
 // - a sub-directory exists in both the source and and the destination and their permissions do not match
-func MoveAllFilesRecursively(fromPath, toPath string, counter progress.Incrementer) *failures.Failure {
+func MoveAllFilesRecursively(fromPath, toPath string, cb MoveAllFilesCallback) *failures.Failure {
 	if !DirExists(fromPath) {
 		return failures.FailOS.New("err_os_not_a_directory", fromPath)
 	} else if !DirExists(toPath) {
@@ -453,32 +455,35 @@ func MoveAllFilesRecursively(fromPath, toPath string, counter progress.Increment
 		subFromPath := filepath.Join(fromPath, fileInfo.Name())
 		subToPath := filepath.Join(toPath, fileInfo.Name())
 		toInfo, err := os.Stat(subToPath)
-		if toPathExists := err == nil; toPathExists {
+		// if stat returns, the destination path exists (either file or directory)
+		toPathExists := err == nil
+		// handle case where destination exists
+		if toPathExists {
 			if fileInfo.IsDir() != toInfo.IsDir() {
 				return failures.FailOS.New("err_incompatible_move_file_dir", subFromPath, subToPath)
 			}
 			if fileInfo.Mode() != toInfo.Mode() {
 				logging.Warning(locale.T("warn_move_incompatible_modes", subFromPath, subToPath))
 			}
-			if toInfo.IsDir() {
-				fail := MoveAllFilesRecursively(subFromPath, subToPath, counter)
-				if fail != nil {
-					return fail
-				}
-				// source path should be empty now
-				err := os.Remove(subFromPath)
-				if err != nil {
-					return failures.FailOS.Wrap(err)
-				}
-				continue
+		}
+		if toPathExists && toInfo.IsDir() {
+			fail := MoveAllFilesRecursively(subFromPath, subToPath, cb)
+			if fail != nil {
+				return fail
 			}
+			// source path should be empty now
+			err := os.Remove(subFromPath)
+			if err != nil {
+				return failures.FailOS.Wrap(err)
+			}
+		} else {
 			logging.Warning(locale.T("warn_move_destination_overwritten", subFromPath))
+			err = os.Rename(subFromPath, subToPath)
+			if err != nil {
+				return failures.FailOS.Wrap(err)
+			}
+			cb()
 		}
-		err = os.Rename(subFromPath, subToPath)
-		if err != nil {
-			return failures.FailOS.Wrap(err)
-		}
-		counter.Increment()
 	}
 	return nil
 }
